refactor(web-todomvc-redwood): migrate to new multi-store API
