+++ examples/web-todomvc-redwood/src/app/todomvc/TodoApp.tsx
@@ -2,19 +2,15 @@
 
 import 'todomvc-app-css/index.css'
 
-import { makePersistedAdapter } from '@livestore/adapter-web'
-import LiveStoreSharedWorker from '@livestore/adapter-web/shared-worker?sharedworker'
-import { LiveStoreProvider } from '@livestore/react'
+import { StoreRegistry, StoreRegistryProvider } from '@livestore/react'
 import { FPSMeter } from '@overengineering/fps-meter'
 import type React from 'react'
-import { unstable_batchedUpdates as batchUpdates } from 'react-dom'
+import { Suspense, useState } from 'react'
 
 import { Footer } from './components/Footer.tsx'
 import { Header } from './components/Header.tsx'
 import { MainSection } from './components/MainSection.tsx'
 import { VersionBadge } from './components/VersionBadge.tsx'
-import { schema } from './livestore/schema.ts'
-import LiveStoreWorker from './livestore.worker.ts?worker'
 
 const AppBody: React.FC = () => (
   <section className="todoapp">
@@ -24,37 +20,18 @@ const AppBody: React.FC = () => (
   </section>
 )
 
-// Check for reset persistence only in browser environment (not during SSR)
-const resetPersistence =
-  typeof window !== 'undefined' &&
-  import.meta.env.DEV &&
-  new URLSearchParams(window.location.search).get('reset') !== null
+export const TodoApp: React.FC = () => {
+  const [storeRegistry] = useState(() => new StoreRegistry())
 
-// Clean up the URL by removing the reset parameter (only in browser)
-if (resetPersistence) {
-  const searchParams = new URLSearchParams(window.location.search)
-  searchParams.delete('reset')
-  window.history.replaceState(null, '', `${window.location.pathname}?${searchParams.toString()}`)
+  return (
+    <Suspense fallback={<div>Loading LiveStore...</div>}>
+      <StoreRegistryProvider storeRegistry={storeRegistry}>
+        <div style={{ top: 0, right: 0, position: 'absolute', background: '#333' }}>
+          <FPSMeter height={40} />
+        </div>
+        <AppBody />
+        <VersionBadge />
+      </StoreRegistryProvider>
+    </Suspense>
+  )
 }
-
-const adapter = makePersistedAdapter({
-  storage: { type: 'opfs' },
-  worker: LiveStoreWorker,
-  sharedWorker: LiveStoreSharedWorker,
-  resetPersistence,
-})
-
-export const TodoApp: React.FC = () => (
-  <LiveStoreProvider
-    schema={schema}
-    adapter={adapter}
-    renderLoading={(_) => <div>Loading LiveStore ({_.stage})...</div>}
-    batchUpdates={batchUpdates}
-  >
-    <div style={{ top: 0, right: 0, position: 'absolute', background: '#333' }}>
-      <FPSMeter height={40} />
-    </div>
-    <AppBody />
-    <VersionBadge />
-  </LiveStoreProvider>
-)
