+++ examples/web-todomvc-redwood/src/app/todomvc/components/Header.tsx
@@ -1,13 +1,13 @@
 'use client'
 
-import { useStore } from '@livestore/react'
 import type React from 'react'
 
 import { uiState$ } from '../livestore/queries.ts'
 import { events } from '../livestore/schema.ts'
+import { useAppStore } from '../livestore/store.ts'
 
 export const Header: React.FC = () => {
-  const { store } = useStore()
+  const store = useAppStore()
   const { newTodoText } = store.useQuery(uiState$)
 
   const updatedNewTodoText = (text: string) => store.commit(events.uiStateSet({ newTodoText: text }))
